% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{contour_plot_fun}
\alias{contour_plot_fun}
\title{Contour plot for showing predicted power}
\usage{
contour_plot_fun(
  combined_data,
  power_estimate,
  cont_breaks,
  multiple_samples = FALSE
)
}
\arguments{
\item{combined_data}{data used for fitting gam}

\item{power_estimate}{predicted power}

\item{cont_breaks}{breaks for contour plot}
}
\value{
ggplot2 object
}
\description{
Contour plot for showing predicted power
}
