% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deseq_fun_est}
\alias{deseq_fun_est}
\title{Title}
\usage{
deseq_fun_est(
  metadata_list,
  countdata_list,
  num_cores = 10,
  ref_name = "control"
)
}
\arguments{
\item{metadata_list}{: list of metadata}

\item{countdata_list}{: list of otu tables}

\item{num_cores}{: number of cores}

\item{ref_name}{reference for fold change calculation}
}
\description{
Title
}
