% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{dispersion_fit}
\alias{dispersion_fit}
\title{Title: fit non-linear model: a  + b/mean_abundance to dispersion estimates}
\usage{
dispersion_fit(dispersion, logmean)
}
\arguments{
\item{logmean}{vector of log mean abundance}

\item{dispersion:}{dispersion estimates from deseq}
}
\value{
: param: parameters for the fit with confidence intervals
}
\description{
Title: fit non-linear model: a  + b/mean_abundance to dispersion estimates
}
