% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dnormmix}
\alias{dnormmix}
\title{Takes a single vector of parameters on the unconstrained scale}
\usage{
dnormmix(x, par, logmean, ..., log = FALSE)
}
\arguments{
\item{par}{softmax(prob), mean, log(sd) on unconstrained scale}

\item{...}{other input calues taken by \code{genmixpars}}

\item{log}{log scale}
}
\description{
Takes a single vector of parameters on the unconstrained scale
}
