% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dnormmix0}
\alias{dnormmix0}
\title{Density function for the mixture of Gaussian distributions}
\usage{
dnormmix0(x, probs, muvals, sdvals, log = FALSE)
}
\arguments{
\item{x}{vector}

\item{probs}{mixture proportions}

\item{muvals}{values of the mean}

\item{sdvals}{values for standard deviataions}

\item{log}{log scale}
}
\value{
likelihood
}
\description{
takes pars as three vectors (prob, mean, sd), on constrained scale
}
