% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_fun}
\alias{filter_fun}
\title{Title}
\usage{
filter_fun(countdata, metadata, abund_thresh = 5, sample_thresh = 3)
}
\arguments{
\item{countdata:}{otu table}

\item{abund_thresh:}{minimum taxa abundance threshold}

\item{sample_thresh:}{minimum sample threshold}
}
\value{
countdata_filt: filtered otu table
control_count:  otu table for control group
treat_count:  otu table for treatment group
}
\description{
Title
}
