% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_low_count}
\alias{filter_low_count}
\title{Filter to remove low abundant taxa}
\usage{
filter_low_count(
  countdata,
  metadata,
  abund_thresh = 5,
  sample_thresh = 3,
  sample_colname,
  group_colname
)
}
\arguments{
\item{countdata}{otu table}

\item{metadata}{dataframe with 2 rows sample names and group names}

\item{abund_thresh}{minimum number of taxa abundance threshold}

\item{sample_thresh}{minimum number of sample threshold}

\item{sample_colname}{column names of the samples}

\item{group_colname}{column names of the groups or conditions}
}
\value{
filtered otu count data
}
\description{
Filter to retain only taxa with at least \code{abund_thresh} counts in at least \code{sample_thresh} samples
}
