% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{gam_fit}
\alias{gam_fit}
\title{Title}
\usage{
gam_fit(
  deseq_est_list,
  true_lfoldchange_list,
  true_lmean_list,
  grid_len = 50,
  alpha_level = 0.1
)
}
\arguments{
\item{deseq_est_list}{a list containing fold change, pvalues and other estimates from \code{DESeq2}}

\item{true_lfoldchange_list}{list containing simulated log fold change used  for simulating the count data}

\item{true_lmean_list}{list containing simulated log mean count used  for simulating the count data}

\item{grid_len}{number of grids for}

\item{alpha_level}{significance level for power calculations}
}
\value{
A list

fit_2d is the fitted scam object

power_estimate predicted power estimated using fit_2d

combined_data tibble containing pvlaues and used for GAM fit
}
\description{
Title
}
