% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genmixpars}
\alias{genmixpars}
\title{generate normal mixture parameters (prob vector, mean vector, sd vector
for a specified set of 'x' values (logmean)}
\usage{
genmixpars(x, pars, np = 2, sd_ord = 2)
}
\arguments{
\item{x}{independent variable}

\item{pars}{parameter vector: first logit-probs (np-1),
then mean parameters (2 per component: intercepts, slopes),
then var parameters (varord + 1 per component: intercepts,
slopes, quad coeffs, etc.)}

\item{np}{number of components in mixture}

\item{sd_ord}{order of logsd model (2 = quadratic)}
}
\description{
generate normal mixture parameters (prob vector, mean vector, sd vector
for a specified set of 'x' values (logmean)
}
