% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{logfoldchange_fit}
\alias{logfoldchange_fit}
\title{Title}
\usage{
logfoldchange_fit(
  logmean,
  logfoldchange,
  ncore = 2,
  max_sd_ord = 2,
  max_np = 5,
  minval = -5,
  maxval = 5,
  itermax = 100,
  NP = 800,
  seed = 100
)
}
\arguments{
\item{seed}{}

\item{logmean:}{vector of log mean abundance}

\item{logfoldchange:}{vector of log fold change}

\item{ncore:}{number of cores to use}

\item{max_sd_ord:}{the maximum order of polynomial function to fit to
standard deviation parameter.
This must be either 1 (linear) or 2(quadratic)}

\item{max_np:}{maximum number of Gaussian components to check for}

\item{minval:}{minimum value for DEoptim search}

\item{maxval:}{maximum value for DEoptim search}
}
\description{
Title
}
