% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{logfoldchange_sim_fun}
\alias{logfoldchange_sim_fun}
\title{Title}
\usage{
logfoldchange_sim_fun(
  logmean_sim,
  logfoldchange_param,
  max_lfc = 15,
  max_iter = 10000
)
}
\arguments{
\item{logmean_sim:}{simulated log mean abundance}

\item{logfoldchange_param:}{list containing the
par = optimal parameters for log fold change fit
np = optimal number of components for log fold change
sd_order: order polynomial for standard deviation parameter
for log fold change}

\item{max_lfc:}{maximum log fold change value to be simulated}

\item{max_iter:}{maximum iteration while simulating logfoldchange < max_lfc}
}
\value{
: lfc: vector of  log fold change values
}
\description{
Title
}
