% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{logmean_fit}
\alias{logmean_fit}
\title{Title}
\usage{
logmean_fit(logmean)
}
\arguments{
\item{logmean:}{vector of log mean abundances}
}
\value{
: components: optimal number of Gaussian components and
logmean_param: parameters for  Gaussian mixtures
density_plot: plot comparing simulation from fit with observation
}
\description{
Title
}
