% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{logmean_sim_fun}
\alias{logmean_sim_fun}
\title{Title}
\usage{
logmean_sim_fun(logmean_param, notu)
}
\arguments{
\item{logmean_param}{log mean parameter}

\item{notu}{number of otus}
}
\description{
Title
}
