% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myrnormmix}
\alias{myrnormmix}
\title{Simulating from a mixture of Gaussian}
\usage{
myrnormmix(par, logmean, ...)
}
\arguments{
\item{par}{parameters (mean, standard deviation and mixture proportion of the mixture of Gaussian)}

\item{logmean}{log mean count of taxa}

\item{...}{other parameters taken by \code{genmixpars}}
}
\value{
random values from a  a mixture of Gaussian
}
\description{
Simulating from a mixture of Gaussian
}
