% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nllfun}
\alias{nllfun}
\title{Objective function}
\usage{
nllfun(par, vals, logmean, np, sd_ord)
}
\arguments{
\item{par}{parameters of the mixture of Gaussian}

\item{vals}{values of log fold change}

\item{logmean}{log mean count}

\item{np}{number of Gaussian components}

\item{sd_ord}{order of polynomial function to model standard deviation parameters
(1 - linear function and 2- quad)}
}
\value{
value for the objective function
}
\description{
Objective function
}
