% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{optimal.comp}
\alias{optimal.comp}
\title{Title: compute the optimal number of components using parametric bootstrap}
\usage{
optimal.comp(logmean, sig = 0.05, max.comp = 4, max.boot = 100)
}
\arguments{
\item{sig:}{significance level}

\item{max.comp:}{maximum number of Gaussian components to check}

\item{max.boot:}{maximum number of bootstraps  to test}
}
\value{
: ncomp: the optimal number of components for log mean abundance
}
\description{
Title: compute the optimal number of components using parametric bootstrap
}
