% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{polyfun}
\alias{polyfun}
\title{General-purpose log-likelihood function, vectorized sum(pars*x^i)}
\usage{
polyfun(pars, x)
}
\arguments{
\item{pars}{parameters}

\item{x}{log mean count}
}
\value{
values representing output for the polynomial fuction (f(x))
}
\description{
General-purpose log-likelihood function, vectorized sum(pars*x^i)
}
