% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.nb-package.R
\docType{package}
\name{power.nb-package}
\alias{power.nb}
\alias{power.nb-package}
\title{power.nb: Power and Sample Size Calculation for Differential Abundance Microbiome Studies}
\description{
Provides functions for estimating statistical power and required sample sizes in differential abundance microbiome studies using negative binomial models. The methods are based on Agronah and Bolker (2025) \doi{10.1371/journal.pone.0318820}. The package includes tools for simulation-based power analysis and sample size estimation using generalized additive models (GAMs), and visualization utilities for exploring the relationship between power, effect size, abundance, and sample size.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://michaelagronah.com/power.nb/}
  \item Report bugs at \url{https://github.com/magronah/power.nb/issues}
}

}
\author{
\strong{Maintainer}: Michael Agronah \email{michaelagronah@gmail.com} (\href{https://orcid.org/0009-0007-2655-4094}{ORCID})

Authors:
\itemize{
  \item Ben Bolker \email{bolker@mcmaster.ca} (\href{https://orcid.org/0000-0002-2127-0443}{ORCID})
}

}
\keyword{internal}
