% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_data}
\alias{read_data}
\title{Extract specified data from a list of datasets}
\usage{
read_data(dataset_list, extract_name)
}
\arguments{
\item{dataset_list}{A list of datasets from which data will be extracted.
Each element of the list is assumed to be a dataset (typically a list or dataframe).}

\item{extract_name}{A string representing the name of the component or column
to be extracted from each dataset in the \code{dataset_list}. The function looks
for this name within each dataset.}
}
\value{
A list containing the extracted data. Each element corresponds to
the extracted component from the datasets in \code{dataset_list}. The names of
the list elements are taken from the names of \code{dataset_list}.
}
\description{
This function extracts a specific component (data) from a list of datasets.
The component to extract is specified by the \code{extract_name} parameter, and
the function returns a list containing the extracted data from each dataset.
}
\examples{
# Example dataset list
dataset1 <- list(countdata = matrix(1:9, nrow = 3), metadata = data.frame(id = 1:3))
dataset2 <- list(countdata = matrix(10:18, nrow = 3), metadata = data.frame(id = 4:6))
dataset_list <- list(dataset1 = dataset1, dataset2 = dataset2)

# Extract 'countdata' from each dataset in the list
result <- read_data(dataset_list, "countdata")
print(result)

}
