% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rnormmix0}
\alias{rnormmix0}
\title{general-purpose normal-mixture deviate generator: takes \emph{matrices}
of probabilities, means, sds}
\usage{
rnormmix0(n, probs, muvals, sdvals)
}
\arguments{
\item{n}{number of observations}

\item{probs}{mixture proportions}

\item{muvals}{values for the mean}

\item{sdvals}{values for the standard deviation}
}
\value{
simulations
}
\description{
general-purpose normal-mixture deviate generator: takes \emph{matrices}
of probabilities, means, sds
}
