% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{countdata_sim_fun}
\alias{countdata_sim_fun}
\title{Title}
\usage{
countdata_sim_fun(
  logmean_param,
  logfoldchange_param,
  dispersion_param,
  nsamp_per_group = NULL,
  ncont = NULL,
  ntreat = NULL,
  notu,
  nsim = 1,
  disp_scale = 0.3,
  max_lfc = 15,
  maxlfc_iter = 1000,
  seed = NULL
)
}
\arguments{
\item{logmean_param}{:  log mean parameters}

\item{logfoldchange_param}{: log fold change parameters}

\item{dispersion_param}{: dispersion parameters}

\item{nsamp_per_group}{: number of samples per group}

\item{notu}{: number of otus}

\item{nsim}{:number of simulations}

\item{disp_scale}{: scale parameter for dispersion}

\item{max_lfc}{: maximum log fold change}

\item{maxlfc_iter}{: maximum number of iterations}

\item{seed}{: seed for simulation}

\item{ncont:}{number of control samples}

\item{ntreat:}{number of treatment samples}
}
\description{
Title
}
