% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deseqfun}
\alias{deseqfun}
\title{Estimate log fold changes using \code{DESeq2}.}
\usage{
deseqfun(
  countdata,
  metadata,
  alpha_level = 0.1,
  ref_name = "NT",
  minReplicatesForReplace = Inf,
  cooksCutoff = TRUE,
  independentFiltering = TRUE,
  shrinkage_method = "normal"
)
}
\arguments{
\item{countdata}{A matrix of OTU count data where rows represent taxa and columns represent samples.}

\item{metadata}{A dataframe containing sample information with two rows: one for sample names and one for group names.}

\item{alpha_level}{The significance level for determining differential expression. Default is 0.1.}

\item{ref_name}{The reference group for calculating fold changes. Default is "NT" (Non-Treatment).}

\item{minReplicatesForReplace}{DESeq2's parameter to control the minimum number of replicates required for replacing outliers during dispersion estimation. Default is \code{Inf} (no replacement).}

\item{cooksCutoff}{DESeq2's parameter for removing outliers based on Cook's distance. Default is \code{TRUE} (outlier removal enabled).}

\item{independentFiltering}{DESeq2's parameter for independent filtering. Default is \code{TRUE}.}

\item{shrinkage_method}{DESeq2's shrinkage method for fold changes. Default is \code{"normal"}. Other options include \code{"apeglm"} or \code{"ashr"}.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{logfoldchange}: A vector of log fold change estimates for each taxa.
\item \code{logmean}: A vector of log mean counts for each taxa (arithmetic mean for taxa across all subjects).
\item \code{dispersion}: A vector of dispersion estimates for each taxa.
\item \code{deseq_estimate}: A dataframe containing DESeq2 results, including \code{baseMean}, \code{log2FoldChange}, \code{lfcSE}, \code{pvalue}, and \code{padj}.
\item \code{normalised_count}: A matrix of normalized count data.
}
}
\description{
This function estimates log fold changes (LFC) for microbiome count data using the DESeq2 package.
It is strongly recommended to keep the following defaults:
\itemize{
\item \code{minReplicatesForReplace=Inf}
\item \code{cooksCutoff=TRUE}
\item \code{independentFiltering=TRUE}
These options are particularly useful when estimating fold changes to fit the mixture of Gaussian distributions.
}
}
\examples{
# Example usage
set.seed(101)
countdata <- matrix(rpois(500, 3), ncol = 10, nrow = 50)
# Simulated OTU count data with 50 taxa and 10 samples

metadata <- data.frame(Samples = paste("Sample", 1:10, sep = "_"),
             Groups = rep(c("Control", "Treatment"), each = 5))

result <- deseqfun(countdata, metadata, ref = "Control"

# Examine the results
result$logfoldchange  # Log fold changes
result$logmean  # Log mean counts
result$dispersion  # Dispersion estimates
result$deseq_estimate  # DESeq2 results
result$normalised_count  # Normalized count data

}
