% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deseqfun}
\alias{deseqfun}
\title{Title}
\usage{
deseqfun(
  countdata,
  metadata,
  alpha_level = 0.1,
  ref_name = "NT",
  minReplicatesForReplace = Inf,
  cooksCutoff = TRUE,
  independentFiltering = TRUE,
  shrinkage_method = "normal"
)
}
\arguments{
\item{ref_name:}{reference for fold change calculation}

\item{minReplicatesForReplace:}{deseq's parameter to control minimum number of
replicates needed for the replacement of outliers during dispersion estimation.}

\item{cooksCutoff:}{deseq's outlier removal or shrinkage}

\item{independentFiltering:}{deseq's independent filtering}

\item{shrinkage_method:}{deseq's shrinkage method}
}
\description{
Title
}
