% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_parnames}
\alias{gen_parnames}
\title{mm}
\usage{
gen_parnames(np, sd_ord)
}
\arguments{
\item{np}{number of Gaussian components}

\item{sd_ord}{order of polynomial function to model standard deviation parameters
(1 - linear function and 2- quadratic function)}
}
\description{
mm
}
