% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{logfoldchange_sim_fun}
\alias{logfoldchange_sim_fun}
\title{Simulate Log Fold Change Values}
\usage{
logfoldchange_sim_fun(
  logmean_sim,
  logfoldchange_param,
  max_lfc = 15,
  max_iter = 10000,
  seed = 121
)
}
\arguments{
\item{logmean_sim}{A numeric vector of simulated log mean abundances.}

\item{logfoldchange_param}{A list containing parameters for the log fold change simulation:
\itemize{
\item \code{par}: Optimal parameters for the log fold change fit.
\item \code{np}: Optimal number of components for the log fold change model.
\item \code{sd_ord}: Order of the polynomial used for the standard deviation parameter of the log fold change.
}}

\item{max_lfc}{A numeric value specifying the maximum allowable absolute log fold change value. Default is 15.}

\item{max_iter}{An integer specifying the maximum number of iterations allowed to ensure all simulated LFC values are within the \code{max_lfc} range. Default is 10,000.}

\item{seed}{random-number seed}
}
\value{
A numeric vector of simulated log fold change values (\code{lfc}).
}
\description{
This function generates simulated log fold change (LFC) values based on the provided log mean abundance and LFC parameters.
The simulation ensures that the generated LFC values remain within a specified maximum range by iterating until convergence or until a maximum iteration limit is reached.
}
\examples{
# Define simulated log mean abundance
logmean_sim <- rnorm(100, mean = 0, sd = 1)

# Define parameters for log fold change simulation
logfoldchange_param <- list(
  par = c(1, -0.5, 0.2), # Example parameters
  np = 2,                # Number of components
  sd_ord = 2             # Order of polynomial for SD
)

# Simulate log fold change values
logfoldchange_sim_fun(
  logmean_sim = logmean_sim,
  logfoldchange_param = logfoldchange_param,
  max_lfc = 10,
  max_iter = 5000
)
}
