% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optimal.comp}
\alias{optimal.comp}
\title{Computes the optimal number of gaussian components for log mean count}
\usage{
optimal.comp(logmean, sig = 0.05, max.comp = 4, max.boot = 100)
}
\arguments{
\item{logmean}{vector of log mean abundances of taxa}

\item{sig}{significance level to compare against p-value}

\item{max.comp}{maximum number of Gaussian components to compare sequentially}

\item{max.boot}{maximum number of bootstraps simulations}
}
\value{
The best number of components for fitting the distribution of log mean abundance
}
\description{
The number of gaussian components is determined using the
using parametric bootstrap
}
\examples{
logmean  = rnorm(100)
optimal.comp(logmean,sig=0.05,max.comp=4,max.boot=100)
}
