% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{power_fun_ss}
\alias{power_fun_ss}
\title{Title}
\usage{
power_fun_ss(
  deseq_est_list,
  true_logfoldchange,
  true_logmean,
  sample_vec,
  alpha = 0.1,
  notu
)
}
\arguments{
\item{alpha}{}

\item{deseq_estimate_list}{list containing simulations for each sample size}
}
\description{
Title
}
