% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_cal.R
\name{uniroot_ss}
\alias{uniroot_ss}
\title{predict.scam seem to have a bug and
would not predict row data with only one row}
\usage{
uniroot_ss(target_power, logmean, abs_lfc, model, xmin, xmax)
}
\description{
predict.scam seem to have a bug and
would not predict row data with only one row
}
